requirements.txt
python-telegram-bot
dotenv
requests